Refactor: 공통 레이아웃 컴포넌트 폴더명 개선
